Electronic lock and IoT access control
An access control system is a solution adopted for several security problems,
and to provide a little more security to users and their properties. Through this
technology, it is possible to extract management knowledge: knowing who is inside
the environment where it was implemented, time and access data, in real time. There
are several access control systems, applied to varied profiles, from simple locks to
facial recognition systems, biometrics, retina and iris, and speech recognition, but at
a high cost. The proposed system aims to improve an existing low-cost access
system, implementing new access features such as the use of passwords captured
through a membrane keyboard, which is an alternative in cases of missing or
forgotten RFID tags or cards, and implementation of online data storage on a server,
providing to the system administrators a more convenient monitoring. The project
described is a microcontrolled access control system implemented with an Arduino
Mega 2560 board. Through this board it is made an identification of the RFID tags or
cards, or the provision of the password through the membrane keypad to authorize
the activation of an electromagnetic lock, thus providing access to the restricted
environment. To provide more flexibility and security, the system relies on an online
software that accesses a MySQL database on a server, which stores all the data of
the users present or not in the restricted area. That is, with this software it is possible
to know in real time, who accessed the controlled locations. Access to these
information is restricted to the system administrator and takes place through a mobile
application for Android platform developed in MIT App Inventor. In this work topics
about the process of development of the system are presented. The tests for the
verification of the operation and comparison of the total cost in relation to systems
offered commercially are exposed at the end.